Renaming utils file to rules
